Apply code review comments
