+++ src/Microsoft.VisualStudio.Threading/TplExtensions.cs
@@ -109,128 +109,6 @@ namespace Microsoft.VisualStudio.Threading
         }
 
         /// <summary>
-        /// Creates a new <see cref="CancellationToken"/> that is canceled when any of a set of other tokens are canceled.
-        /// </summary>
-        /// <param name="original">The first token.</param>
-        /// <param name="other">The second token.</param>
-        /// <returns>A struct that contains the combined <see cref="CancellationToken"/> and a means to release memory when you're done using it.</returns>
-        public static CombinedCancellationToken CombineWith(this CancellationToken original, CancellationToken other)
-        {
-            if (original.IsCancellationRequested)
-            {
-                return new CombinedCancellationToken(original);
-            }
-
-            if (other.IsCancellationRequested)
-            {
-                return new CombinedCancellationToken(other);
-            }
-
-            if (original.CanBeCanceled && other.CanBeCanceled)
-            {
-                // This is the most expensive path to take since it involves allocating memory and requiring disposal.
-                // Before this point we've checked every condition that would allow us to avoid it.
-                return new CombinedCancellationToken(CancellationTokenSource.CreateLinkedTokenSource(original, other));
-            }
-
-            return original.CanBeCanceled ? new CombinedCancellationToken(original)
-                : new CombinedCancellationToken(other);
-        }
-
-        /// <summary>
-        /// Creates a new <see cref="CancellationToken"/> that is canceled when any of a set of other tokens are canceled.
-        /// </summary>
-        /// <param name="original">The first token.</param>
-        /// <param name="others">The additional tokens.</param>
-        /// <returns>A struct that contains the combined <see cref="CancellationToken"/> and a means to release memory when you're done using it.</returns>
-        public static CombinedCancellationToken CombineWith(this CancellationToken original, params CancellationToken[] others)
-        {
-            Requires.NotNull(others, nameof(others));
-
-            if (original.IsCancellationRequested)
-            {
-                return new CombinedCancellationToken(original);
-            }
-
-            int cancelableTokensCount = original.CanBeCanceled ? 1 : 0;
-            foreach (var other in others)
-            {
-                if (other.IsCancellationRequested)
-                {
-                    return new CombinedCancellationToken(other);
-                }
-
-                if (other.CanBeCanceled)
-                {
-                    cancelableTokensCount++;
-                }
-            }
-
-            switch (cancelableTokensCount)
-            {
-                case 0:
-                    return new CombinedCancellationToken(CancellationToken.None);
-                case 1:
-                    if (original.CanBeCanceled)
-                    {
-                        return new CombinedCancellationToken(original);
-                    }
-
-                    foreach (var other in others)
-                    {
-                        if (other.CanBeCanceled)
-                        {
-                            return new CombinedCancellationToken(other);
-                        }
-                    }
-
-                    throw Assumes.NotReachable();
-                case 2:
-                    CancellationToken first = CancellationToken.None;
-                    CancellationToken second = CancellationToken.None;
-
-                    if (original.CanBeCanceled)
-                    {
-                        first = original;
-                    }
-
-                    foreach (var other in others)
-                    {
-                        if (other.CanBeCanceled)
-                        {
-                            if (first.CanBeCanceled)
-                            {
-                                second = other;
-                            }
-                            else
-                            {
-                                first = other;
-                            }
-                        }
-                    }
-
-                    Assumes.True(first.CanBeCanceled && second.CanBeCanceled);
-
-                    // Call the overload that takes two CancellationTokens explicitly to avoid an array allocation.
-                    return new CombinedCancellationToken(CancellationTokenSource.CreateLinkedTokenSource(first, second));
-                default:
-                    // This is the most expensive path to take since it involves allocating memory and requiring disposal.
-                    // Before this point we've checked every condition that would allow us to avoid it.
-                    var cancelableTokens = new CancellationToken[cancelableTokensCount];
-                    int i = 0;
-                    foreach (var other in others)
-                    {
-                        if (other.CanBeCanceled)
-                        {
-                            cancelableTokens[i++] = other;
-                        }
-                    }
-
-                    return new CombinedCancellationToken(CancellationTokenSource.CreateLinkedTokenSource(cancelableTokens));
-            }
-        }
-
-        /// <summary>
         /// Applies one task's results to another.
         /// </summary>
         /// <typeparam name="T">The type of value returned by a task.</typeparam>
@@ -880,53 +758,6 @@ namespace Microsoft.VisualStudio.Threading
             public void GetResult()
             {
                 // Never throw here.
-            }
-        }
-
-        /// <summary>
-        /// Provides access to a <see cref="System.Threading.CancellationToken"/> that combines multiple other tokens,
-        /// and allows convenient disposal of any applicable <see cref="CancellationTokenSource"/>.
-        /// </summary>
-        public struct CombinedCancellationToken : IDisposable
-        {
-            /// <summary>
-            /// The object to dispose when this struct is disposed.
-            /// </summary>
-            private readonly CancellationTokenSource cts;
-
-            /// <summary>
-            /// Initializes a new instance of the <see cref="CombinedCancellationToken"/> struct
-            /// that contains an aggregate <see cref="System.Threading.CancellationToken"/> whose source must be disposed.
-            /// </summary>
-            /// <param name="cancellationTokenSource">The cancellation token source.</param>
-            public CombinedCancellationToken(CancellationTokenSource cancellationTokenSource)
-            {
-                this.cts = cancellationTokenSource;
-                this.Token = cancellationTokenSource.Token;
-            }
-
-            /// <summary>
-            /// Initializes a new instance of the <see cref="CombinedCancellationToken"/> struct
-            /// that represents just a single, non-disposable <see cref="System.Threading.CancellationToken"/>.
-            /// </summary>
-            /// <param name="cancellationToken">The cancellation token</param>
-            public CombinedCancellationToken(CancellationToken cancellationToken)
-            {
-                this.cts = null;
-                this.Token = cancellationToken;
-            }
-
-            /// <summary>
-            /// Gets the combined cancellation token.
-            /// </summary>
-            public CancellationToken Token { get; }
-
-            /// <summary>
-            /// Disposes the <see cref="CancellationTokenSource"/> behind this combined token, if any.
-            /// </summary>
-            public void Dispose()
-            {
-                this.cts?.Dispose();
             }
         }
